### 2018-06-07 #223 @bnasr
xROI
Delineate Region of Interests (ROI's) and Extract Time-Series Data from Digital Repeat

```
Package: xROI
Title: Delineate Region of Interests (ROI's) and Extract Time-Series Data from Digital Repeat Photography Images
Version: 0.1.2
Author: Bijan Seyednasrollah, Thomas Milliman, Andrew D. Richardson
Maintainer: Bijan Seyednasrollah <bijan.s.nasr@gmail.com>
Description: Digital repeat photography and near-surface remote sensing have been used by environmental scientists to study the environmental change for nearly a decade. However, a user-friendly, reliable, and robust platform to extract color-based statistics and time-series from a large stack of images is still lacking. Here, we present an interactive open-source toolkit, called 'xROI', that facilitate the process time-series extraction and improve the quality of the final data. 'xROI' provides a responsive environment for scientists to interactively a) delineate regions of interest (ROI), b) handle field of view (FOV) shifts, and c) extract and export time series data characterizing image color (i.e. red, green and blue channel digital numbers for the defined ROI). Using 'xROI', user can detect FOV shifts without minimal difficulty. The software gives user the opportunity to readjust the mask files or redraw new ones every time an FOV shift occurs. 'xROI' helps to significantly improve data accuracy and continuity.
    raster,
    rgdal,
    sp,
    jpeg,
    tiff,
    shiny,
    shinyjs,
    shinyBS,
    shinyAce,
    shinyTime,
    shinyFiles,
    shinydashboard,
    shinythemes,
    colourpicker,
    rjson,
    plotly,
    moments,
    RCurl,
    graphics,
    stats
RoxygenNote: 6.0.1.9000

```

https://github.com/bnasr/xROI

[e.g., "data extraction, because the package parses a scientific data file format"]
data extraction, because the package extract color based statistics from stack of digital repeat images.

Environmental scientist who use digital repeat photography to study the environmental changes such as vegetation phenology, salt marsh restoration, monitoring tidal wetlands, assessing gross primary productivity, investigating growth in croplands and anyone who is interested to extract time series data from digital images.

No.

0 error, 0 warning, 0 notes submitted on CRAN

name of the package is not all in lower case: xROI
