### 2018-06-11 #223 @sckott
thanks for your submission @bnasr !

## programmatic work

A high level issue is the interactive Shiny interface. We don't have a policy in our documents yet as we haven't had a submission before that's purely a Shiny app, but we are in the process of adding it now. Basically, Shiny interfaces are great, but we'd like there to be a way for users to do the same thing in whole or at least in part, programatically. One way to do this is have a code tab in your Shiny app which gives users the code to do what they did in the Shiny app, and eventually they may learn how to use your pkg through code as well as via Shiny. One example of combining interactive with code is tabulizer, see https://github.com/ropensci/tabulizer#code-examples

Do you think you'd be willing to try to incorporate functions/etc. for using your pkg programmatically? Or are you committed to only the Shiny interface?

## tests

As per our [packaging guide](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md#-testing) we do require every submission to have a test suite.  Please add a test suite - check out this guide https://rstudio.github.io/shinytest/articles/shinytest.html for how to add tests for shiny apps.
