### 2018-06-18 #223 @bnasr
Thanks @sckott for your review and comments and sorry for the delay. I've been working on this.
The package actually does have some low-level programmatic work options in addition to the high-level shiny interface. Part of this has been documented in the package, however not in the vignette page. I will try to add this to the vignette and prepare a test case. Thanks.
