### 2019-02-25 #223 @sckott
@bnasr be aware that our policies https://ropensci.github.io/dev_guide/policies.html#review-process state that this issue will be closed after one year from applying the holding label (jul 2019) and you'll need to resubmit
