### 2025-05-12 #223 @mpadge
@Rajesh78079 The repository source is https://github.com/bbcrown/xROI, so you'd have to ask over there. But note that it hasn't been touched for years, so you should expect that the package will no longer work.
